introduce usecases
