+++ pages/autoform.js
@@ -1,10 +1,11 @@
 import React from 'react';
-import { Container } from 'semantic-ui-react';
+import { Container, Header } from 'semantic-ui-react';
 import { AutoForm } from 'uniforms-semantic';
 import { createSchemaBridge } from './_uniforms';
 import { withRouter } from 'next/router';
+import Registration from '../usecases/Registration';
 
-function AutoFormPage({ router, request, schema }) {
+function AutoFormPage({ router, request, response, schema }) {
   console.log({ request });
 
   function onSubmit(query) {
@@ -19,34 +20,18 @@ function AutoFormPage({ router, request, schema }) {
         onSubmit={onSubmit}
         showInlineError={true}
       />
+      {response.greeting && <Header as="h1">{response.greeting}</Header>}
     </Container>
   );
 }
 export default withRouter(AutoFormPage);
 
-export async function getServerSideProps({ query }) {
-  const schema = {
-    title: 'Guest',
-    type: 'object',
-    properties: {
-      firstName: { type: 'string', minLength: 1 },
-      lastName: { type: 'string', minLength: 1 },
-      workExperience: {
-        description: 'Work experience in years',
-        type: 'integer',
-        minimum: 0,
-        maximum: 100
-      },
-      zip: { type: 'string', pattern: '[0-9]{5}' },
-      color: { type: 'string', enum: ['red', 'amber', 'green'] }
-    },
-    required: ['firstName', 'lastName']
-  };
-  const request = query;
-
+export async function getServerSideProps({ query: request }) {
+  const usecase = new Registration();
   const props = {
-    schema,
-    request
+    request,
+    response: await usecase.process(request),
+    schema: usecase.getSchema(request)
   };
   return { props };
 }
